platform/requirements.txt
# Plataforma LMS (certificados, BD, conteudo encriptado). Python 3.11+
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
sqlalchemy>=2.0.36,<3
python-multipart>=0.0.12,<1
pyjwt[crypto]>=2.9.0,<3
bcrypt>=4.2.0,<5
cryptography>=43.0.0,<46
fpdf2>=2.8.0,<3
pydantic-settings>=2.6.0,<3
email-validator>=2.2.0,<3
